RND-14 : automatically calculate areas that need to be redacted
This code automatically calculates the PdfCleanUpLocations for pdfSweep,
based on a regular expression provided by the end-user. The code allows
various options (using font information to determine the height of the
boxes vs. taking the minimum viable height, setting a color, as well as
a composite pattern to merge multiple easier regular expressions into
one). This code does not use the graphics state of TextRenderInfo
objects beyond the validity of the state.