Preserve image cs while cleaning it up and processing fixes.
After clean image in case of not JPEG there were an additional compression parameter set with LZW value for tiff images.
As this if statement has been never invoked even in "tiff-lzw" tests, then we can remove it.
Changed the parsing of JPEG images to javax always as apache has some problems with it.
Added logic with not copying the original cs in case when the new one is the same.
Added tests.

DEVSIX-1908